requirements.txt
anyio==3.6.2
astroid==2.14.1
beautifulsoup4==4.11.2
black==23.1.0
click==8.1.3
deta 
dill==0.3.6
deta==1.1.0
fastapi==0.89.1
feedparser==6.0.10
h11==0.14.0
httptools==0.5.0
httpx==0.24.1
idna==3.4
isort==5.12.0
joblib==1.2.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
mypy-extensions==0.4.3
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
platformdirs==2.6.2
pydantic==1.10.4
pylint==2.16.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
sentry-sdk ==1.21.1 
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
typing-extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
wrapt==1.14.1
